WIP: This adds support to the putdata script for dumping out the RDF contents of a wikibase and then re-ingesting it through the WDQS munge and loadData scripts.
This was tried out with 0.3.6 and 0.3.97 but with the same results.

When loading the data there are some distinct differences between the
resulting dataset.

The imported export has additonal dump triples saying that the data was imported.

The http://wikiba.se/ontology#timestamp triples are also missing from each entity.

{
	"s": "http://tophats.wbaas.localhost/entity/Q3",
	"p": "http://wikiba.se/ontology#timestamp",
	"o": "2022-02-02T10:08:24Z"
},